Add support for username-based history.

+++ build-workflow.sh
@@ -1,29 +1,35 @@
-#!/bin/bash -e
+#!/usr/bin/env zsh
+
+wffiles=(icon.png info.plist README.md LICENCE.txt alfssh)
 
 here="$( cd "$( dirname "$0" )"; pwd )"
-
 
 log() {
     echo "$@" > /dev/stderr
 }
 
+cleanup() {
+    local p="${here}/build"
+    log "Cleaning up ..."
+    test -d "$p" && rm -rf ./build/
+}
+
 pushd "$here" &> /dev/null
 
-# Get metadata from info.plist
-# bundleid="$( /usr/libexec/PlistBuddy -c 'Print :bundleid' info.plist )"
-#
-# if test -z "$bundleid"; then
-#     log "No bundle ID found in info.plist"
-#     exit 1
-# fi
-#
-# name="$( /usr/libexec/PlistBuddy -c 'Print :name' info.plist )"
-#
-# if test -z "$name"; then
-#     log "No workflow name found in info.plist"
-#     exit 1
-# fi
+log "Building executable(s) ..."
+go build -v -o ./alfssh ./cmd/alfssh
+ST_BUILD=$?
+if [ "$ST_BUILD" != 0 ]; then
+    log "Error building executable."
+    cleanup
+    popd &> /dev/null
+    exit $ST_BUILD
+fi
 
+chmod 755 ./alfssh
+# cp -v ./alfssh ./build/alfssh
+
+log
 
 log "Cleaning ./build ..."
 rm -rvf ./build
@@ -34,30 +40,17 @@ log "Copying assets to ./build ..."
 
 mkdir -vp ./build
 
-cp -v icon.png ./build/
-cp -v info.plist ./build/
-cp -v README.md ./build/
-cp -v LICENCE.txt ./build/
+for n in $wffiles; do
+    ln -v "$n" ./build/
+done
 
 log
 
-log "Building executable(s) ..."
-go build -v -o ./alfssh
-ST_BUILD=$?
-if [ "$ST_BUILD" != 0 ]; then
-    log "Error building executable."
-    rm -rf ./build/
-    popd &> /dev/null
-    exit $ST_BUILD
-fi
-
-chmod 755 ./alfssh
-cp -v ./alfssh ./build/alfssh
-
 # Get the dist filename from the executable
-zipfile="$(./alfssh --distname 2> /dev/null)"
+zipfile="$(./alfssh print distname 2> /dev/null)"
 if [ "$?" -ne 0 ]; then
     log "Error getting distname from alfssh."
+    cleanup
     exit 1
 fi
 
@@ -79,7 +72,8 @@ zip "../${zipfile}" ./*
 ST_ZIP=$?
 if [ "$ST_ZIP" != 0 ]; then
     log "Error creating .alfredworkflow file."
-    rm -rf ./build/
+    popd &> /dev/null
+    cleanup
     popd &> /dev/null
     exit $ST_ZIP
 fi
@@ -87,8 +81,7 @@ popd &> /dev/null
 
 log
 
-log "Cleaning up ..."
-rm -rvf ./build/
+cleanup
 
 log "Wrote '${zipfile}' in '$( pwd )'"
